+++ Idp.Swiyu.IdentityProvider/HostingExtensions.cs
@@ -60,9 +60,7 @@ internal static class HostingExtensions
 
         builder.Services.AddHttpClient();
         builder.Services.AddOptions();
-        // Add services to the container.
         builder.Services.AddRazorPages();
-        builder.Services.AddControllers();
 
         builder.Services.AddDbContext<ApplicationDbContext>(options =>
             options.UseSqlite(builder.Configuration.GetConnectionString("DefaultConnection")));
@@ -111,8 +109,6 @@ internal static class HostingExtensions
 
         app.MapRazorPages()
             .RequireAuthorization();
-
-        app.MapControllers();
 
         return app;
     }