+++ Idp.Swiyu.IdentityProvider/Pages/Account/Manage/IdentityAlreadyVerified.cshtml.cs
@@ -1,0 +1,11 @@
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.RazorPages;
+
+namespace Idp.Swiyu.IdentityProvider.Pages.Account.Manage;
+
+public class IdentityAlreadyVerifiedModel : PageModel
+{
+    public void OnGet()
+    {
+    }
+}